+++ src/Lmc.ServiceIdentification/SimpleTypes.fs
@@ -2,18 +2,27 @@ namespace Lmc.ServiceIdentification
 
 [<RequireQualifiedAccess>]
 module Domain =
+    let [<Literal>] Pattern = @"^([a-zA-Z]+)$"
+    let parseStrict = SimpleType.parseStrict Pattern Domain DomainError.Empty DomainError.InvalidFormat
+
     let value (Domain domain) = domain
     let map = SimpleType.map value Domain
     let lower = map String.toLower
 
 [<RequireQualifiedAccess>]
 module Context =
+    let [<Literal>] Pattern = @"^([a-zA-Z]+)$"
+    let parseStrict = SimpleType.parseStrict Pattern Context ContextError.Empty ContextError.InvalidFormat
+
     let value (Context context) = context
     let map = SimpleType.map value Context
     let lower = map String.toLower
 
 [<RequireQualifiedAccess>]
 module Purpose =
+    let [<Literal>] Pattern = @"^([a-zA-Z]+)$"
+    let parseStrict = SimpleType.parseStrict Pattern Purpose PurposeError.Empty PurposeError.InvalidFormat
+
     let value (Purpose purpose) = purpose
     let map = SimpleType.map value Purpose
     let lower = map String.toLower
@@ -26,6 +35,9 @@ module PurposePattern =
 
 [<RequireQualifiedAccess>]
 module Version =
+    let [<Literal>] Pattern = @"^([a-zA-Z]+[a-zA-Z0-9]*)$"
+    let parseStrict = SimpleType.parseStrict Pattern Version VersionError.Empty VersionError.InvalidFormat
+
     let value (Version version) = version
     let map = SimpleType.map value Version
     let lower = map String.toLower
@@ -38,6 +50,9 @@ module VersionPattern =
 
 [<RequireQualifiedAccess>]
 module Zone =
+    let [<Literal>] Pattern = @"^([a-zA-Z]+)$"
+    let parseStrict = SimpleType.parseStrict Pattern Zone ZoneError.Empty ZoneError.InvalidFormat
+
     let value (Zone zone) = zone
     let map = SimpleType.map value Zone
     let lower = map String.toLower
@@ -50,6 +65,9 @@ module ZonePattern =
 
 [<RequireQualifiedAccess>]
 module Bucket =
+    let [<Literal>] Pattern = @"^([a-zA-Z]+)$"
+    let parseStrict = SimpleType.parseStrict Pattern Bucket BucketError.Empty BucketError.InvalidFormat
+
     let value (Bucket bucket) = bucket
     let map = SimpleType.map value Bucket
     let lower = map String.toLower