+++ CHANGELOG.md
@@ -17,6 +17,13 @@
     - `Instance.lower`
     - `Box.lower`
     - `ServiceIdentification.lower`
+- Add `parseStrict` function to *SimpleTypes* and a specific *ErrorTypes*:
+    - `Domain.parseStrict`
+    - `Context.parseStrict`
+    - `Purpose.parseStrict`
+    - `Version.parseStrict`
+    - `Zone.parseStrict`
+    - `Bucket.parseStrict`
 
 ## 5.1.1 - 2021-02-09
 - Rename `Lmc.ServiceIdentification.fsproj` back to `ServiceIdentification.fsproj` to match old library